LDA analysis.
"results"  folder to leave LDA conclusions. df_step3.json.bz2 to read some quotes
Given a quote dataset, analyze according to LDA methodology, which are the most repeated topics, which are the words inside that topic and their weight inside the topic, and the visualization of the distribution of words in general in the dataset (blue), the distribution of the words inside a topic (red, lambda = 1), which are the most "unique" words that define a topic in the upper part of the right graph if lambda = 0. And also, on the left graph, we see the topics, the ones that are further away are the most different.